feat(search): Phase 2 — Trie + Inverted Index + TF-IDF search engine
Implements a complete document search engine from scratch.

TextProcessor:
- O(n) tokenization with single-pass character loop
- Static stop word set using unordered_set for O(1) lookup
- Careful unsigned char cast before tolower (avoids UB)
- Separate tokenizeQuery() preserves stop words in user queries

Trie:
- Sparse node design: unordered_map<char, unique_ptr<TrieNode>>
  trades fixed 208-byte children array for actual-child-only storage
- O(m) insert/contains/startsWith (m = word length)
- DFS collectWords for prefix completion
- Const/non-const findNode using const_cast pattern to avoid duplication
- unique_ptr chains ensure recursive RAII destruction

InvertedIndex:
- unordered_map<string, vector<Posting>> for O(1) word lookup
- Posting struct: docId + termFreq + firstPosition (12 bytes, no padding)
- Static empty posting list returned by const-ref for missing words
- Linear scan for existing postings (cache-friendly for short lists)

SearchEngine:
- Smoothed TF-IDF: log((1+N)/(1+df)) + 1 prevents zero scores
- Score accumulation via unordered_map<DocumentID, SearchResult>
- std::sort with lambda comparator for descending rank
- Trie used only for prefix search; InvertedIndex handles ranking

Demonstrates: data structures, algorithmic complexity, cache locality,
const-correctness, STL internals, information retrieval algorithms.

+++ src/pdf/PdfWriter.cpp
@@ -163,8 +163,8 @@ void PdfWriter::write(
             stream.writeStr("<< /Type /Page\n");
             stream.writeF("   /Parent %d 0 R\n", pagesNum);
             stream.writeF("   /MediaBox [0 0 %.2f %.2f]\n",
-                          page.width(), page.height());
-            stream.writeF("   /Contents %d 0 R\n", alloc.contentObjNum);
+static_cast<double>(page.width()), static_cast<double>(page.height()));           
+ stream.writeF("   /Contents %d 0 R\n", alloc.contentObjNum);
 
             // Font resources dictionary
             if (!fontsCopy.empty()) {
